requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0
pyyaml>=6.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-pptx>=0.6.21

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0
pytesseract>=0.3.10

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0
whisper-openai>=1.0.0
speechrecognition>=3.8.1

# Machine Learning & NLP
scikit-learn>=1.1.0
transformers>=4.20.0
sentence-transformers>=2.2.0
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Vector databases and search
faiss-cpu>=1.7.3
chromadb>=0.4.0
pinecone-client>=2.2.0

# LLM integration
langchain>=0.1.0
langchain-community>=0.0.10
ollama>=0.1.7

# Utilities
tqdm>=4.64.0
click>=8.0.0
rich>=12.0.0
loguru>=0.6.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: GPU support (uncomment if using CUDA)
# faiss-gpu>=1.7.3
# torch>=1.12.0+cu116
